Provide actual files and folders, add removing observers when needed
