+++ Magellan/Classes/Modules/LocationDetails/Presenter/LocationDetailsPresenter.swift
@@ -7,69 +7,6 @@ import UIKit
 import MessageUI
 import SafariServices
 
-final class LocationDetailsReviewablePresenter {
-
-    weak var view: LocationDetailsViewProtocol?
-    let decorated: LocationDetailsDecorable
-    let localizator: LocalizedResourcesFactoryProtocol
-    let place: PlaceInfo
-    let dataSource: PlaceReviewDataSource
-
-    internal init(place: PlaceInfo,
-                  decorated: LocationDetailsDecorable,
-                  localizator: LocalizedResourcesFactoryProtocol,
-                  dataSource: PlaceReviewDataSource) {
-        self.place = place
-        self.decorated = decorated
-        self.localizator = localizator
-        self.dataSource = dataSource
-    }
-    
-    func setupContent() {
-        decorated.setupContent()
-        
-        var decoratedItems = decorated.items
-        let items = dataSource.apply(.init(place: place)).map {
-            LocationSectionViewModel(title: nil, header: $0, items: $0.items)
-        }
-        decoratedItems.append(contentsOf: items)
-        decorated.setItems(decoratedItems)
-        
-        view?.reload()
-    }
-    
-}
-extension LocationDetailsReviewablePresenter: LocationDetailsPresenterProtocol {
-    
-    var delegate: LocationDetailsPresenterDelegate? {
-        get { decorated.delegate }
-        set { decorated.delegate = newValue }
-    }
-    var items: [LocationSectionViewModel] { decorated.items }
-    
-    func dismiss() {
-        decorated.dismiss()
-    }
-    
-    func viewDidLoad() {
-        setupContent()
-    }
-}
-
-protocol LocationDetailsDecorable: LocationDetailsPresenterProtocol {
-            
-    func setItems(_ items: [LocationSectionViewModel])
-    
-    func setupContent()
-}
-
-extension LocationDetailsPresenter: LocationDetailsDecorable {
-    
-    func setItems(_ items: [LocationSectionViewModel]) {
-        self.items = items
-    }
-
-}
 
 final class LocationDetailsPresenter {
     
@@ -84,7 +21,9 @@ final class LocationDetailsPresenter {
         place = PlaceInfoViewModel(place: placeInfo)
         formatter = phoneFormatter
         self.localizator = localizedResourcesFactory
-        
+    }
+    
+    func becomeObserver() {
         NotificationCenter.default.addObserver(self,
                                                selector: #selector(localizationChanged),
                                                name: .init(rawValue: localizator.notificationName),
@@ -93,6 +32,10 @@ final class LocationDetailsPresenter {
     
     @objc func localizationChanged() {
         setupContent()
+    }
+    
+    deinit {
+        NotificationCenter.default.removeObserver(self)
     }
     
     func setupContent() {
@@ -175,3 +118,10 @@ extension LocationDetailsPresenter: LocationDetailsPresenterProtocol {
 }
 
 extension LocationDetailsPresenter: UrlHandlerProtocol { }
+extension LocationDetailsPresenter: LocationDetailsDecorable {
+    
+    func setItems(_ items: [LocationSectionViewModel]) {
+        self.items = items
+    }
+
+}
